Fix review comments

+++ tests/integration/data/BackendFactoryTests.cpp
@@ -56,7 +56,7 @@ protected:
          ConfigValue{ConfigType::Double}.defaultValue(0.5)},
         {"database.cassandra.max_request_retry_delay",
          ConfigValue{ConfigType::Double}.defaultValue(5.0)},
-        {"database.cassandra.request_timeout", ConfigValue{ConfigType::Integer}.optional()},
+        {"database.cassandra.request_timeout", ConfigValue{ConfigType::Double}.optional()},
         {"database.cassandra.username", ConfigValue{ConfigType::String}.optional()},
         {"database.cassandra.password", ConfigValue{ConfigType::String}.optional()},
         {"database.cassandra.certfile", ConfigValue{ConfigType::String}.optional()},
